+++ src/components/Navigation/SideNavigation/ContextMenu/Item/Item.stories.tsx
@@ -4,7 +4,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { Item } from "./Item";
 
 export default {
-  title: "Side Navigation/Item",
+  title: "Navigation/Side Navigation/Context Menu/Item",
   component: Item,
   tags: ["autodocs"],
 } as Meta<typeof Item>;
@@ -13,5 +13,6 @@ export const Default: StoryObj<typeof Item> = {
   args: {
     icon: EnvelopeOpenIcon,
     text: "Hello World",
+    keyboard: "⌘+B",
   },
 };
