+++ src/components/Navigation/SideNavigation/Item/Item.stories.tsx
@@ -4,7 +4,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 import { Item } from "./Item";
 
 export default {
-  title: "Side Navigation/Context Menu/Item",
+  title: "Navigation/Side Navigation/Item",
   component: Item,
   tags: ["autodocs"],
 } as Meta<typeof Item>;
@@ -13,6 +13,5 @@ export const Default: StoryObj<typeof Item> = {
   args: {
     icon: EnvelopeOpenIcon,
     text: "Hello World",
-    keyboard: "⌘+B",
   },
 };
